+++ src/main/java/com/dev3/app/WeChatAppApplication.java
@@ -1,4 +1,4 @@
-package com.volvo.it.vgas;
+package com.dev3.app;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
