+++ src/main/java/com/dev3/app/repositories/TextMessageRepository.java
@@ -1,0 +1,9 @@
+package com.dev3.app.repositories;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.dev3.app.entity.TextMessage;
+
+public interface TextMessageRepository extends JpaRepository<TextMessage,Long>{
+
+}
