+++ src/main/java/com/dev3/app/web/WeChatMessageUtil.java
@@ -1,0 +1,68 @@
+package com.dev3.app.web;
+
+import java.io.IOException;
+import java.io.InputStream;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+
+import javax.servlet.http.HttpServletRequest;
+
+import org.dom4j.Document;
+import org.dom4j.DocumentException;
+import org.dom4j.Element;
+import org.dom4j.io.SAXReader;
+
+import com.dev3.app.entity.TextMessage;
+import com.thoughtworks.xstream.XStream;
+
+public class WeChatMessageUtil {
+    /**
+     * convert xml to Map Collection
+     * 
+     * @param request
+     * @return
+     */
+    public static Map<String, String> xmlToMap(HttpServletRequest request) {
+        Map<String, String> map = new HashMap<String, String>();
+        SAXReader reader = new SAXReader();
+        InputStream ins = null;
+        try {
+            ins = request.getInputStream();
+        } catch (IOException e1) {
+            e1.printStackTrace();
+        }
+        Document doc = null;
+        try {
+            doc = reader.read(ins);
+        } catch (DocumentException e1) {
+            e1.printStackTrace();
+        }
+        Element root = doc.getRootElement();
+        @SuppressWarnings("unchecked")
+        List<Element> list = root.elements();
+        for (Element e : list) {
+            map.put(e.getName(), e.getText());
+        }
+        try {
+            ins.close();
+        } catch (IOException e1) {
+            e1.printStackTrace();
+        }
+        return map;
+    }
+
+    /**
+     * Serialize TextMessage Object To XML
+     * 
+     * @param textMessage
+     * @return
+     */
+    public static String textMessageToXml(TextMessage textMessage) {
+        XStream xstream = new XStream();
+        xstream.alias("xml", textMessage.getClass());
+        return xstream.toXML(textMessage);
+
+    }
+
+}
