+++ src/main/java/com/dev3/app/web/WeChatTokenInfo.java
@@ -1,0 +1,40 @@
+package com.dev3.app.web;
+
+public class WeChatTokenInfo {
+	public String getSignature() {
+		return signature;
+	}
+
+	public void setSignature(String signature) {
+		this.signature = signature;
+	}
+
+	public String getTimestamp() {
+		return timestamp;
+	}
+
+	public void setTimestamp(String timestamp) {
+		this.timestamp = timestamp;
+	}
+
+	public String getNonce() {
+		return nonce;
+	}
+
+	public void setNonce(String nonce) {
+		this.nonce = nonce;
+	}
+
+	public String getEchostr() {
+		return echostr;
+	}
+
+	public void setEchostr(String echostr) {
+		this.echostr = echostr;
+	}
+
+	private String signature;
+	private String timestamp;
+	private String nonce;
+	private String echostr;
+}
