+++ src/main/java/com/dev3/app/web/package-info.java
@@ -1,0 +1,2 @@
+
+package com.dev3.app.web;
