--- src/main/java/com/volvo/it/vgas/entities/package-info.java
@@ -1,8 +1,0 @@
-/**
- * 
- */
-/**
- * @author a032049
- *
- */
-package com.volvo.it.vgas.entities;
