--- src/main/java/com/volvo/it/vgas/webapi/WeChatTokenInfo.java
@@ -1,40 +1,0 @@
-package com.volvo.it.vgas.webapi;
-
-public class WeChatTokenInfo {
-	public String getSignature() {
-		return signature;
-	}
-
-	public void setSignature(String signature) {
-		this.signature = signature;
-	}
-
-	public String getTimestamp() {
-		return timestamp;
-	}
-
-	public void setTimestamp(String timestamp) {
-		this.timestamp = timestamp;
-	}
-
-	public String getNonce() {
-		return nonce;
-	}
-
-	public void setNonce(String nonce) {
-		this.nonce = nonce;
-	}
-
-	public String getEchostr() {
-		return echostr;
-	}
-
-	public void setEchostr(String echostr) {
-		this.echostr = echostr;
-	}
-
-	private String signature;
-	private String timestamp;
-	private String nonce;
-	private String echostr;
-}
