--- src/main/java/com/volvo/it/vgas/webapi/package-info.java
@@ -1,8 +1,0 @@
-/**
- * 
- */
-/**
- * @author a032049
- *
- */
-package com.volvo.it.vgas.webapi;
